Commit
The “STUDENT MANAGEMENT SYSTEM” is a web based application system
that provides the users to store and retrieves the student details in an easy and efficient way.
The Administration store new information and delete unwanted information from anywhere
and at any time which in turn saves time. Also this web based system provides a good
security features to all the users since we can avoid piracy in handling information. The
creation and management of accurate, up-to-date information regarding a students’ academic
career is critically important in the university as well as colleges. Student information system
deals with all kind of student details, academic related reports.